refactor: Refactor file paths for NavigationBar and SideNavigationBar components to match new folder structure
