+++ app/(sub-page)/components/side-navigation-bar.tsx
@@ -1,6 +1,13 @@
 'use client';
 import { HamburgerMenuIcon } from '@radix-ui/react-icons';
-import { Sheet, SheetTrigger, SheetContent, SheetHeader, SheetTitle, SheetDescription } from './sheet/sheet';
+import {
+  Sheet,
+  SheetTrigger,
+  SheetContent,
+  SheetHeader,
+  SheetTitle,
+  SheetDescription,
+} from '../../ui/view/molecule/sheet/sheet';
 
 export default function SideNavigationBar() {
   return (
